# apps/console/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/drivers": {
      "get": {
        "operationId": "DriversController_listDrivers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all active drivers (basic info only, HOS fetched from external API)",
        "tags": [
          "Drivers"
        ]
      },
      "post": {
        "operationId": "DriversController_createDriver",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new driver (basic info only)",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/v1/drivers/{driver_id}": {
      "put": {
        "operationId": "DriversController_updateDriver",
        "parameters": [
          {
            "name": "driver_id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update driver basic info",
        "tags": [
          "Drivers"
        ]
      },
      "delete": {
        "operationId": "DriversController_deleteDriver",
        "parameters": [
          {
            "name": "driver_id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete driver (set isActive=false)",
        "tags": [
          "Drivers"
        ]
      },
      "get": {
        "operationId": "DriversController_getDriver",
        "parameters": [
          {
            "name": "driver_id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver details"
          },
          "404": {
            "description": "Driver not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get driver by ID",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/v1/drivers/{driverId}/hos": {
      "get": {
        "operationId": "DriversController_getDriverHOS",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get live HOS data for driver from integration (with cache fallback)",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/v1/drivers/pending/list": {
      "get": {
        "operationId": "DriversController_getPendingDrivers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all pending drivers awaiting activation",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/v1/drivers/inactive/list": {
      "get": {
        "operationId": "DriversController_getInactiveDrivers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all inactive drivers",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/v1/drivers/{driver_id}/activate": {
      "post": {
        "operationId": "DriversController_activateDriver",
        "parameters": [
          {
            "name": "driver_id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate a pending driver",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/v1/drivers/{driver_id}/deactivate": {
      "post": {
        "operationId": "DriversController_deactivateDriver",
        "parameters": [
          {
            "name": "driver_id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate an active driver",
        "tags": [
          "Drivers"
        ]
      }
    },
    "/api/v1/drivers/{driver_id}/reactivate": {
      "post": {
        "operationId": "DriversController_reactivateDriver",
        "parameters": [
          {
            "name": "driver_id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivate an inactive driver",
        "tags": [
          "Drivers"
        ]
      }
    },
...
    "/api/v1/external/tms/drivers/{driverId}": {
      "get": {
        "operationId": "ExternalApisController_getTmsDriver",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "description": "Driver ID (e.g., TMS-D001)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get single driver from mock Project44 TMS API",
        "tags": [
          "External API Mock"
        ]
      }
    },
...
  },
  "info": {
    "title": "SALLY API",
    "description": "Fleet Operations Assistant API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://sally-api.apps.appshore.in/api/v1",
      "description": "Staging"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "type": "http",
        "description": "Enter your API key (sk_staging_...)",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateDriverDto": {
        "type": "object",
        "properties": {
          "driver_id": {
            "type": "string",
            "example": "DRV-001",
            "description": "Unique driver identifier"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Driver full name"
          }
        },
        "required": [
          "driver_id",
          "name"
        ]
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Smith",
            "description": "Driver full name"
          }
        }
      },
...
    }
  }
}